Move device test in own file
